VSR: Send commit_min in do_view_change message
Instead of sending commit_max, which is not strictly required.

Sending commit_min enables the new primary to determine whether
it must trust and therefore repair using another replica's op.

For example, in situations where the hash chain has a gap that
would otherwise prevent the repair, leading to a log fork.